api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
asyncpg==0.30.0
pyjwt[crypto]==2.10.1
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.9.1
mcp==1.27.0
turbopuffer==0.5.1
mistune==3.1.3
python-multipart==0.0.20
